requirements.txt
Pillow
selenium
pyautogui